Refactor routes for minions resource into minions-router.js
